# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tracking daily activities, streaks, and analytics.\n\n## Features\n- User authentication with JWT tokens\n- Activity tracking with customizable types\n- Streak tracking to maintain consistency\n- Weekly analytics with daily breakdowns\n- Privacy controls for user profiles\n- Tile configuration for dashboard customization\n\n## Rate Limits\n- Login/Register: 5 requests per minute per IP\n- Password Reset: 3 requests per minute per IP\n- General API: 60 requests per minute per user",
        "title": "Growth Tracker API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/aman1117/growth-tracker"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/tile-config": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get dashboard tile configuration for authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tile Config"
                ],
                "summary": "Get tile configuration",
                "responses": {
                    "200": {
                        "description": "Tile configuration",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.TileConfigResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Save dashboard tile configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tile Config"
                ],
                "summary": "Save tile configuration",
                "parameters": [
                    {
                        "description": "Tile configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.SaveTileConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Config saved",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tile-config/user": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get dashboard tile configuration for another user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tile Config"
                ],
                "summary": "Get tile config by username",
                "parameters": [
                    {
                        "description": "Target username",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.GetTileConfigByUsernameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tile configuration",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.TileConfigResponse"
                        }
                    },
                    "400": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Private account",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_aman1117_backend_internal_dto.ErrorResponse": {
            "description": "Generic error response",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "error_code": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.GetTileConfigByUsernameRequest": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.SaveTileConfigRequest": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/models.JSONB"
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.SuccessResponse": {
            "description": "Generic success response",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_aman1117_backend_internal_dto.TileConfigResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.JSONB"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.JSONB": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}